requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
google-generativeai
sentence-transformers
tiktoken
matplotlib
pandas